Add Router to App and create component pages
