+++ package.json
@@ -9,6 +9,7 @@
     "react-dom": "^16.4.0",
     "react-emotion": "^9.1.3",
     "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4",
     "redux": "^4.0.0",
     "storybook": "^1.0.0",
