+++ src/components/History.js
@@ -1,0 +1,7 @@
+import React, { Component } from 'react'
+
+export default class History extends Component {
+  render() {
+    return <div>History </div>
+  }
+}
