+++ src/styles/Grid.js
@@ -1,0 +1,8 @@
+import styled from 'react-emotion'
+
+const Grid = styled('div')`
+  display: grid;
+  grid-template-rows: auto;
+  height: 100vh;
+`
+export default Grid
